Diabetes- On-Set-Detection
It predicts if the person has diabetes or not depending on certain parameters like glucose, blood pressure, skin thickness etc

On-set-Diabetes-detection We are using some parameters like pregnacy,glucose,blood pressure etc to predict if the person has diabetes or not. This project works on PIMA Indians Diabetes Data Set. It uses the potential of deep learning to predict the on-set of diabetes in females. The concept of hyperparameter tuning using grid search helps in training the model and predict the output. The advantage of this project is that it takes into account general human body parameters instead of medical parametes obtained using medical test. The model obtained a 78.2% accuracy score.